requirements.txt
absl-py==1.4.0
Bottleneck 
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.42.1
grpcio==1.56.0
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.6.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib 
kiwisolver==1.4.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.2
mkl-fft==1.3.6
mkl-random 
mkl-service==2.4.0
mpmath==1.2.1
networkx==3.0
numexpr 
numpy==1.21.5
packaging==23.1
pandas==1.4.3
Pillow==9.3.0
protobuf==3.20.0
pyparsing==3.0.9
python-dateutil 
pytorch-metric-learning==2.2.0
pytz 
requests==2.28.1
scikit-learn 
scipy==1.10.1
seaborn==0.12.2
six 
sklearn==0.0.post5
sympy==1.11.1
tensorboard==1.15.0
threadpoolctl 
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.3.6
zipp==3.15.0
